third_party/blink/web_tests/fast/dom/tabindex-clamp-expected.txt

This page tests that the tabindex attribute is accepted for values between -2147483648 and 2147483647. Values outside of this range will make the tabIndex reflected value zero.

Testing input elements in the page
getAttribute("tabindex") = 3147483648; tabIndex = 0
getAttribute("tabindex") = 2147483648; tabIndex = 0
getAttribute("tabindex") = 2147483647; tabIndex = 2147483647
getAttribute("tabindex") = 2147483646; tabIndex = 2147483646
getAttribute("tabindex") = 0; tabIndex = 0
getAttribute("tabindex") = -1; tabIndex = -1
getAttribute("tabindex") = -2147483647; tabIndex = -2147483647
getAttribute("tabindex") = -2147483648; tabIndex = -2147483648
getAttribute("tabindex") = -2147483649; tabIndex = 0
getAttribute("tabindex") = -3147483648; tabIndex = 0
Testing setAttribute on an anchor element made with document.createElement
getAttribute("tabindex") = 3147483648; tabIndex = 0
getAttribute("tabindex") = 2147483648; tabIndex = 0
getAttribute("tabindex") = 2147483647; tabIndex = 2147483647
getAttribute("tabindex") = 2147483646; tabIndex = 2147483646
getAttribute("tabindex") = 0; tabIndex = 0
getAttribute("tabindex") = -1; tabIndex = -1
getAttribute("tabindex") = -2147483647; tabIndex = -2147483647
getAttribute("tabindex") = -2147483648; tabIndex = -2147483648
getAttribute("tabindex") = -3147483648; tabIndex = 0
getAttribute("tabindex") = ; tabIndex = 0
getAttribute("tabindex") = foo; tabIndex = 0
Testing tabIndex on an area element made with document.createElement
getAttribute("tabindex") = -1147483648; tabIndex = -1147483648
getAttribute("tabindex") = -2147483648; tabIndex = -2147483648
getAttribute("tabindex") = 2147483647; tabIndex = 2147483647
getAttribute("tabindex") = 2147483646; tabIndex = 2147483646
getAttribute("tabindex") = 0; tabIndex = 0
getAttribute("tabindex") = -1; tabIndex = -1
getAttribute("tabindex") = -2147483647; tabIndex = -2147483647
getAttribute("tabindex") = -2147483648; tabIndex = -2147483648
getAttribute("tabindex") = 1147483648; tabIndex = 1147483648
getAttribute("tabindex") = 0; tabIndex = 0
getAttribute("tabindex") = 0; tabIndex = 0
